Finished Chapter 3 Exercises
I have completed all the exercises in Think Python chapter 3. Though I did not do it the same way as done in the answer key, it does the same thing.  I could have made it even shorter with loops, but tried to only use the things I learned in chapters 1-3